fix: circular dependencies (#83)
