+++ src/ResizeObserverController.ts
@@ -1,4 +1,4 @@
-import { scheduler } from './utils/scheduler';
+import { scheduler, updateCount } from './utils/scheduler';
 
 import { ResizeObserver } from './ResizeObserver';
 import { ResizeObservation } from './ResizeObservation';
@@ -6,21 +6,9 @@ import { ResizeObserverDetail } from './ResizeObserverDetail';
 import { ResizeObserverCallback } from './ResizeObserverCallback';
 import { ResizeObserverOptions } from './ResizeObserverOptions';
 
-import { hasActiveObservations } from './algorithms/hasActiveObservations';
-import { hasSkippedObservations } from './algorithms/hasSkippedObservations';
-import { deliverResizeLoopError } from './algorithms/deliverResizeLoopError';
-import { broadcastActiveObservations } from './algorithms/broadcastActiveObservations';
-import { gatherActiveObservationsAtDepth } from './algorithms/gatherActiveObservationsAtDepth';
+import { resizeObservers } from './utils/resizeObservers';
 
-const resizeObservers: ResizeObserverDetail[] = [];
-const observerMap = new Map();
-let watching = 0;
-
-const updateCount = (n: number): void => {
-  !watching && n > 0 && scheduler.start();
-  watching += n;
-  !watching && scheduler.stop();
-}
+const observerMap = new Map<ResizeObserver, ResizeObserverDetail>();
 
 // Helper to find the correct ResizeObservation, based on a target.
 const getObservationIndex = (observationTargets: ResizeObservation[], target: Element): number => {
@@ -30,23 +18,6 @@ const getObservationIndex = (observationTargets: ResizeObservation[], target: El
     }
   }
   return -1;
-}
-
-/**
- * Runs through the algorithms and
- * broadcasts and changes that are returned.
- */
-const process = (): boolean => {
-  let depth = 0;
-  gatherActiveObservationsAtDepth(depth);
-  while (hasActiveObservations()) {
-    depth = broadcastActiveObservations();
-    gatherActiveObservationsAtDepth(depth);
-  }
-  if (hasSkippedObservations()) {
-    deliverResizeLoopError();
-  }
-  return depth > 0;
 }
 
 /**
@@ -92,6 +63,4 @@ class ResizeObserverController {
   }
 }
 
-const isWatching = (): boolean => !!watching;
-
-export { ResizeObserverController, resizeObservers, process, isWatching };
+export { ResizeObserverController };
