+++ src/algorithms/hasActiveObservations.ts
@@ -1,4 +1,4 @@
-import { resizeObservers } from '../ResizeObserverController';
+import { resizeObservers } from '../utils/resizeObservers';
 import { ResizeObserverDetail } from '../ResizeObserverDetail';
 
 /**
