+++ src/utils/resizeObservers.ts
@@ -1,0 +1,5 @@
+import { ResizeObserverDetail } from '../ResizeObserverDetail';
+
+const resizeObservers: ResizeObserverDetail[] = [];
+
+export { resizeObservers };
